bpf: tests: condense external IP test
Run the IPv4 and IPv6 test from a single file.
