+++ bpf/tests/tc_lb_external_ips.h
@@ -6,17 +6,20 @@
 #include "common.h"
 #include "pktgen.h"
 
-#define ENABLE_IPV4 1
-#define ENABLE_NODEPORT 1
+#define ENABLE_IPV4		1
+#define ENABLE_IPV6		1
+#define ENABLE_NODEPORT		1
 
 #ifdef NORTH_SOUTH_TEST
 # define CLIENT_IP		v4_ext_one
+# define CLIENT_IPV6		v6_ext_node_one
 # define HOOK			netdev_receive_packet
 
 #include "lib/bpf_host.h"
 #elif defined(EAST_WEST_TEST)
 # define ENABLE_SOCKET_LB_HOST_ONLY	1
 # define CLIENT_IP			v4_pod_two
+# define CLIENT_IPV6			v6_pod_two
 # define HOOK				pod_send_packet
 
 #include "lib/bpf_lxc.h"
@@ -27,12 +30,15 @@
 #define CLIENT_PORT		tcp_src_one
 
 #define FRONTEND_IP		v4_ext_two
+#define FRONTEND_IPV6		v6_ext_node_two
 #define FRONTEND_PORT		tcp_svc_one
 
 #define BACKEND_IP		v4_pod_one
+#define BACKEND_IPV6		v6_pod_one
 #define BACKEND_PORT		tcp_dst_one
 
 #define NAT_REV_INDEX		1
+#define NAT_REV_INDEX_2		1
 #define BACKEND_COUNT		1
 #define BACKEND_IFINDEX		11
 #define BACKEND_ID		124
@@ -69,6 +75,30 @@ const __u8 lb4_ew_external_ip[] = {
 	SCAPY_BUF_BYTES(lb4_ew_external_ip)
 };
 
+/* Expected CT tuple to check for conntrack */
+#define EXPECTED_CT_TUPLE6 {\
+	.saddr = *(union v6addr *)CLIENT_IPV6, \
+	.sport = FRONTEND_PORT, \
+	.daddr = *(union v6addr *)FRONTEND_IPV6, \
+	.dport = CLIENT_PORT, \
+	.nexthdr = IPPROTO_TCP, \
+	.flags = TUPLE_F_SERVICE }
+
+/* packet defined in ./scapy/lb_pkt_defs.py */
+const __u8 lb6_ns_external_ip[] = {
+	SCAPY_BUF_BYTES(lb6_ns_external_ip)
+};
+
+/* packet defined in ./scapy/lb_pkt_defs.py */
+const __u8 lb6_ns_external_ip_post_dnat[] = {
+	SCAPY_BUF_BYTES(lb6_ns_external_ip_post_dnat)
+};
+
+/* packet defined in ./scapy/lb_pkt_defs.py */
+const __u8 lb6_ew_external_ip[] = {
+	SCAPY_BUF_BYTES(lb6_ew_external_ip)
+};
+
 /* Test that a request for an ExternalIP service is handled correctly:
  * - For North-South traffic, the packet should be accepted and DNATed to the
  *   backend, and a conntrack entry should be created for the connection.
@@ -102,8 +132,8 @@ int lb4_external_ips_setup(struct __ctx_buff *ctx)
 			      (__u8 *)mac_one, (__u8 *)mac_two);
 	lb_v4_add_service_with_flags(FRONTEND_IP, FRONTEND_PORT,
 				     IPPROTO_TCP, BACKEND_COUNT, NAT_REV_INDEX,
-				     SVC_FLAG_ROUTABLE | SVC_FLAG_EXTERNAL_IP | SVC_FLAG_EXT_LOCAL_SCOPE,
-				     SVC_FLAG_TWO_SCOPES);
+				     SVC_FLAG_ROUTABLE | SVC_FLAG_EXTERNAL_IP |
+				     SVC_FLAG_EXT_LOCAL_SCOPE, SVC_FLAG_TWO_SCOPES);
 	lb_v4_add_backend(FRONTEND_IP, FRONTEND_PORT, BACKEND_COUNT, BACKEND_ID,
 			  BACKEND_IP, BACKEND_PORT, IPPROTO_TCP, 0);
 
@@ -119,8 +149,6 @@ int lb4_external_ips_check(__maybe_unused const struct __ctx_buff *ctx)
 	struct ipv4_ct_tuple tuple = EXPECTED_CT_TUPLE;
 
 	test_init();
-
-	endpoint_v4_del_entry(BACKEND_IP);
 
 	data = ctx_data(ctx);
 	data_end = ctx_data_end(ctx);
@@ -160,6 +188,101 @@ int lb4_external_ips_check(__maybe_unused const struct __ctx_buff *ctx)
 
 	/* Ensure CT entry is not populated (SVC). */
 	ct_entry = map_lookup_elem(get_ct_map4(&tuple), &tuple);
+	assert(!ct_entry);
+#endif
+
+	test_finish();
+}
+
+/* Test that a request for an ExternalIP service is handled correctly:
+ * - For North-South traffic, the packet should be accepted and DNATed to the
+ *   backend, and a conntrack entry should be created for the connection.
+ * - For East-West traffic, the packet should be dropped and no conntrack entry
+ *   should be created for the connection.
+ */
+PKTGEN("tc", "tc_lb6_external_ips")
+int lb6_external_ips_pktgen(struct __ctx_buff *ctx)
+{
+	struct pktgen builder;
+
+	pktgen__init(&builder, ctx);
+
+#ifdef NORTH_SOUTH_TEST
+	scapy_push_data(&builder, lb6_ns_external_ip, sizeof(lb6_ns_external_ip));
+#else
+	scapy_push_data(&builder, lb6_ew_external_ip, sizeof(lb6_ew_external_ip));
+#endif
+
+	pktgen__finish(&builder);
+
+	return 0;
+}
+
+SETUP("tc", "tc_lb6_external_ips")
+int lb6_external_ips_setup(struct __ctx_buff *ctx)
+{
+	policy_add_egress_allow_all_entry();
+
+	endpoint_v6_add_entry((union v6addr *)BACKEND_IPV6, BACKEND_IFINDEX, 0, 0, 0,
+			      (__u8 *)mac_one, (__u8 *)mac_two);
+	lb_v6_add_service_with_flags((union v6addr *)FRONTEND_IPV6, FRONTEND_PORT,
+				     IPPROTO_TCP, BACKEND_COUNT, NAT_REV_INDEX_2,
+				     SVC_FLAG_ROUTABLE | SVC_FLAG_EXTERNAL_IP |
+				     SVC_FLAG_EXT_LOCAL_SCOPE, SVC_FLAG_TWO_SCOPES);
+	lb_v6_add_backend((union v6addr *)FRONTEND_IPV6, FRONTEND_PORT, BACKEND_COUNT, BACKEND_ID,
+			  (union v6addr *)BACKEND_IPV6, BACKEND_PORT, IPPROTO_TCP, 0);
+
+	return HOOK(ctx);
+}
+
+CHECK("tc", "tc_lb6_external_ips")
+int lb6_external_ips_check(__maybe_unused const struct __ctx_buff *ctx)
+{
+	void *data, *data_end;
+	__u32 *status_code;
+	struct ct_entry *ct_entry;
+	struct ipv6_ct_tuple tuple = EXPECTED_CT_TUPLE6;
+
+	test_init();
+
+	data = ctx_data(ctx);
+	data_end = ctx_data_end(ctx);
+	status_code = data;
+
+	assert(data + sizeof(__u32) <= data_end)
+
+#ifdef NORTH_SOUTH_TEST
+	/* Ensure packet is accepted. */
+	assert(*status_code == CTX_ACT_OK);
+
+	/* Ensure packet is correctly DNATed. */
+	ASSERT_CTX_BUF_OFF("lb6_ns_external_ip_post_dnat", "Ether", ctx, sizeof(__u32),
+			   lb6_ns_external_ip_post_dnat, sizeof(lb6_ns_external_ip_post_dnat));
+
+	/* Ensure CT entry is updated accordingly (SVC). */
+	ct_entry = map_lookup_elem(get_ct_map6(&tuple), &tuple);
+	assert(ct_entry);
+	assert(ct_entry->packets == 1);
+	assert(ct_entry->bytes == sizeof(lb6_ns_external_ip_post_dnat));
+#else
+	/* Ensure packet is dropped. */
+	assert(*status_code == CTX_ACT_DROP);
+
+	/* Ensure packet is not DNATed. */
+	ASSERT_CTX_BUF_OFF("lb6_ew_external_ip", "Ether", ctx, sizeof(__u32),
+			   lb6_ew_external_ip, sizeof(lb6_ew_external_ip));
+
+	/* Ensure drop metrics are updated accordingly. */
+	struct metrics_key metric_key = {
+		.reason = -DROP_NO_SERVICE,
+		.dir = METRIC_EGRESS,
+	};
+	__u64 count = 2;
+
+	assert_metrics_count(metric_key, count);
+
+	/* Ensure CT entry is not populated (SVC). */
+	ct_entry = map_lookup_elem(get_ct_map6(&tuple), &tuple);
 	assert(!ct_entry);
 #endif
