+++ bpf/tests/tc_lb_external_ips_ns.c
@@ -3,4 +3,4 @@
 
 #define NORTH_SOUTH_TEST 1
 
-#include "tc_lb4_external_ips.h"
+#include "tc_lb_external_ips.h"
